chore(ultragoal): L0 residual gate + L1 truth + hyperframes_ops size split
Claim-audit C1/M1 as verify-only pass; residual maturity matrix, sound stage
DAG, fixture freeze, and critical-path recompute. Split hyperframes_ops helpers
under 800 LOC with guardrail locks. Docs truth punch list and claim-ledger freeze
for the kinocut-full-build ultragoal portfolio.
