+++ docs/status/2026-07-14-1.8-release-notes.md
@@ -42,8 +42,9 @@ Native MCPB remains staged/local-only and is not published as an external artifa
 
 S13 host joins and S14 dual-class benchmark receipts may be complete in-tree; **S15
 adversarial acceptance was STOP at release.** Full-episode Sonic World is not claimed
-as shipped in 1.8.0.
-
+as shipped in 1.8.0. *(Living residual classes after 1.13.x: see
+[`2026-08-12-residual-maturity-matrix.md`](2026-08-12-residual-maturity-matrix.md) —
+do not use this 1.8.0 boundary as greenfield staffing truth.)*
 ### Install
 
 ```bash
