+++ docs/status/2026-07-14-post-1.8-program-status.md
@@ -1,6 +1,13 @@
 # Kinocut post-1.8 program status
 
-**Current release:** Kinocut 1.8.0 is published as of 2026-07-14. The [GitHub
+> **Evidence snapshot (2026-07-14), not living residual truth.** For post-1.13.2
+> residual staffing and sound honesty, use
+> [`2026-08-12-residual-maturity-matrix.md`](2026-08-12-residual-maturity-matrix.md),
+> [sound residual DAG](2026-08-12-sound-residual-stage-dag.md), and
+> [L1 truth pass](2026-08-12-l1-truth-pass.md). S13–S15 language below is historical
+> relative to the 1.8.0 cut — packages and residual classes have moved since.
+
+**Current release (at time of this note):** Kinocut 1.8.0 is published as of 2026-07-14. The [GitHub
 Release](https://github.com/KyaniteLabs/kinocut/releases/tag/v1.8.0), PyPI, npm, and
 MCP Registry resolve to 1.8.0. The published surface is 142 MCP tools and 121 CLI
 commands.
