+++ docs/status/2026-08-12-l1-truth-pass.md
@@ -1,0 +1,71 @@
+# L1 truth pass (G002 / L1.2) — 2026-08-12
+
+**Ultragoal:** `kinocut-full-build` · story `G002-l1-truth`  
+**Scope:** docs honesty only — **no** `docs/public_claims.json` bump  
+**Depends on:** L0 complete (claim-audit, residual matrix, sound DAG, fixture freeze, critical-path)
+
+## L1.2 false-done punch list
+
+| Named false-done | Correction | Status |
+|------------------|------------|--------|
+| ROADMAP WP-C size claim | Modules ≤800 on tip incl. `hyperframes_ops`→helpers; WP-C **Done** | Closed |
+| ROADMAP WP-D long functions | 0 funcs >80 on tip; WP-D **Done** | Closed |
+| HUMAN_GATES C1/M1 open language | Claim-audit **verify-only pass** 2026-08-12; reopen only with failing case | Closed in `docs/HUMAN_GATES.md` + `.omx/state/l0-claim-audit.md` |
+| PHASE residual accuracy | Phase 3/4/E = deepen residual with modules on tip, not “missing scaffold”; sound section + residual matrix links | Closed in `docs/status/PHASE_CHECKPOINTS.md` |
+| Sound honesty (thin S12 vs packages/S14) | Public join thin S12; S4–S13 packages exist; synthetic S14 ≠ product complete; residual re-run/deepen | Closed in ROADMAP + residual matrix + sound DAG |
+| Stale “S5–S15 incomplete” staffing | July handoffs historical; residual matrix is living authority | Supersession notes on July sound handoff + ROADMAP archive note |
+
+## Claim-ledger rule (L1.3 narrative; freeze until L3)
+
+| Field | Value |
+|-------|--------|
+| Claim owner | Ultragoal ledger leader for `kinocut-full-build` (product may override in writing) |
+| Frozen file | `docs/public_claims.json` |
+| Who may bump | **Only** claim owner, at **L3** claim PR (or product-authorized release cut) |
+| Who must not | Residual L1/L2 agents, opportunistic “sync tip counts,” non-owner PRs |
+
+Recorded in [`ROADMAP.md`](../../ROADMAP.md) (Claim ledger freeze) and this receipt.
+Hard CI enforcement for non-owner bumps remains optional follow-on; **process freeze is active**.
+
+## S+ / site stamp (L1.1 preferred)
+
+| Check | Result |
+|-------|--------|
+| Preferred target | README overall ≥95; site `llms.txt` stamp current |
+| Hard portfolio blocker? | **No** (PRD L1.1 preferred) |
+| Live score this pass | See “S+ attempt” section below |
+
+## Human gates (do not invent)
+
+| Gate | Agent status |
+|------|----------------|
+| #88 directories | Open — external reviews still pending (not agent-closed) |
+| #90 launch moments | Open |
+| #92 first-10 users | Open |
+| #3 Renovate host token | Open |
+
+## S+ attempt (2026-08-12)
+
+| Item | Result |
+|------|--------|
+| Tooling present on host | Yes — `~/.agents/bin/s_plus_engine.py`, `verify-readme-splus` → `verify_readme_splus.py` |
+| Live score this pass | **Skipped (executor shell unavailable)** — could not run `python3 ~/.agents/bin/s_plus_engine.py score …` or `verify-readme-splus` from this worker surface |
+| Last recorded floor | 2026-08-07 excellence audit: fleet **42/42 OK**; Kinocut overall **~89.3–92.6** (misses included `seo_keywords_early`, `geo_entity_def`; sometimes `seo_toc`) |
+| README structure now | `<!-- s-plus-geo -->` marker present; TOC present; entity/TL;DR/FAQ/audience/status/agent surface blocks present |
+| L1.1 status | **Open preferred** (≥95 + site `llms.txt` stamp) — **not** hard portfolio GO; not claimed green at ≥95 |
+
+Parent/orchestrator may re-run:
+
+```bash
+python3 ~/.agents/bin/s_plus_engine.py score KyaniteLabs/kinocut
+# or: score local path per engine CLI
+verify-readme-splus 2>&1 | grep -iE 'kinocut|ok=|fail='
+```
+
+## Authority chain for residuals
+
+1. [`2026-08-12-residual-maturity-matrix.md`](2026-08-12-residual-maturity-matrix.md)  
+2. [`2026-08-12-sound-residual-stage-dag.md`](2026-08-12-sound-residual-stage-dag.md)  
+3. [`2026-08-12-sound-fixture-freeze.md`](2026-08-12-sound-fixture-freeze.md)  
+4. [`.omx/state/critical-path.md`](../../.omx/state/critical-path.md)  
+5. July status files = evidence only, not staffing truth
