+++ docs/status/2026-08-12-sound-residual-rerun.md
@@ -1,0 +1,18 @@
+# Sound residual re-run evidence (G003 start)
+
+**Date:** 2026-08-12  
+**Command:** `python3 -m pytest tests/ -q --tb=no -k "sound or kinocut_sound"`  
+**Result:** **580 passed**, 4355 deselected  
+
+## Implication for residual-CP-S
+
+Stage packages + host joins + S14 synthetic evidence + large green sound test surface → primary agent work is **re-verify/deepen product honesty**, not greenfield S5–S15 rebuild.
+
+## Still open for product full-episode claim
+
+- Live dual-class S14 re-run on required hardware (or `external_host_unavailable`)  
+- S15 acceptance + L3 claim owner  
+- Human independent sound adversarial review residual  
+- ROADMAP must not say only “thin S12 join” without residual matrix pointer  
+
+See: `docs/status/2026-08-12-sound-residual-stage-dag.md`, fixture freeze doc.
